+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fixed an issue with topbar component where clicking anchor links does not close the open menu.
 - Fixed an issue with datepicker not displaying properly, or at all, when clicking on label.
 - Fixed an issue where alert close button got properties from the new `<a>` styling (blue border bottom).
+- Fixed a minor issue with toast not resetting its type, if called without a type after another toast with a type has already been called.
 
 ### Removed